+++ lib/game_team.rb
@@ -8,7 +8,7 @@ class GameTeam
               :goals,
               :shots,
               :tackles
- 
+
   def initialize(data)
     @game_id = data["game_id"]
     @team_id = data["team_id"]
@@ -26,19 +26,5 @@ class GameTeam
     @giveaways = data["giveaways"]
     @takeaways = data["takeaways"]
   end
-
-  # def highest_total_score
-  #   grouped_by_game = @game_teams.group_by do |row|
-  #     row["game_id"]
-  #   end
-  #   game_sum_goals = []
-  #   grouped_by_game.each_pair do |key, value|
-  #     first_team_goals = value.first["goals"].to_i
-  #     second_team_goals = value.last["goals"].to_i
-  #     game_sum_goals << first_team_goals + second_team_goals
-  #   end
-  #   game_sum_goals.max
-  # end
-
 
 end
